backend/pyproject.toml
[tool.poetry]
name = "sqlindex-backend"
version = "0.1.0"
description = "Database RAG & Analytics Platform - Backend API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"
sqlmodel = "^0.0.14"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
qdrant-client = "^1.7.0"
sentence-transformers = "^2.3.0"
langchain = "^0.1.0"
langchain-community = "^0.0.13"
langgraph = "^0.0.20"
ollama = "^0.1.8"
redis = "^5.0.1"
celery = "^5.3.6"
sse-starlette = "^1.8.2"
tenacity = "^8.2.3"
cryptography = "^41.0.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
httpx = "^0.27.0"
ruff = "^0.1.11"
mypy = "^1.8.0"
black = "^23.12.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
